requirements.txt
#python==3
keras>=2.2.0
tensorflow==2.13.0
keras-applications>=1.0.7,<=1.0.8
image-classifiers==1.0.*
efficientnet==1.0.*
matplotlib ==3.7.2
segmentation-models==1.0.1
scikit-learn==1.3.0
opencv-python==4.8.0.76
wandb==0.15.8

#dlib==19.23.1
einops==0.4.1
Flask==2.1.1
#mc==0.0
natsort==8.1.0
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
PyYAML==6.0
requests==2.27.1
scikit_image==0.21.0 #latest
scipy==1.8.0
setuptools==58.0.4
torch==2.0.1 #latest
torchvision== 0.15.2 #latest
tqdm==4.66.1

catalyst==22.4
